(PC-34195)[API] refactor: simplify earliest eligible registration detection
